@@ -5,31 +5,29 @@
   "scripts": {
     "dev": "vitepress dev",
     "build": "vitepress build && esno scripts/build.ts",
+    "preview": "npm run build && vitepress serve",
     "serve": "npm run build && serve .vitepress/dist"
   },
   "dependencies": {
-    "@vueuse/core": "^7.2.2",
-    "@vueuse/shared": "^7.2.2",
-    "vue": "^3.2.24",
-    "vue-global-api": "^0.4.1"
+    "@vueuse/core": "^9.2.0",
+    "@vueuse/shared": "^9.2.0",
+    "vue": "^3.2.38"
   },
   "devDependencies": {
-    "@iconify/json": "^2.0.8",
-    "@rollup/plugin-replace": "^3.0.0",
     "@types/fs-extra": "^9.0.13",
-    "@vueuse/core": "^7.2.2",
-    "esbuild-register": "^3.2.0",
-    "eslint": "^8.4.1",
-    "fast-glob": "^3.2.7",
-    "fs-extra": "^10.0.0",
-    "https-localhost": "^4.7.0",
-    "prism-theme-vars": "^0.2.2",
-    "typescript": "^4.5.2",
-    "unplugin-icons": "^0.12.23",
-    "unplugin-vue-components": "^0.17.6",
-    "vite": "^2.7.1",
+    "@vitejs/plugin-vue": "^3.1.0",
+    "@vueuse/core": "^9.2.0",
+    "esbuild-register": "^3.3.3",
+    "eslint": "^8.23.0",
+    "fast-glob": "^3.2.11",
+    "fs-extra": "^10.1.0",
+    "https-localhost": "^4.7.1",
+    "typescript": "^4.8.2",
+    "unocss": "^0.45.18",
+    "unplugin-vue-components": "^0.22.4",
+    "vite": "^3.1.0",
     "vite-plugin-pwa": "workspace:*",
-    "vite-plugin-windicss": "^1.5.4",
-    "vitepress": "^0.20.4"
+    "vitepress": "^1.0.0-alpha.13",
+    "workbox-window": "^6.5.4"
   }
 }
