@@ -20,32 +23,48 @@
     "run-build-sw": "rimraf dist && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW=true vite build",
     "run-build-sw-reloadsw": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true RELOAD_SW=true SW=true vite build",
     "run-build-sw-claims": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true CLAIMS=true SW=true vite build",
-    "run-build-sw-reloadsw-claims": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true CLAIMS=true RELOAD_SW=true SW=true vite build",
+    "run-build-sw-claims-reloadsw": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true CLAIMS=true RELOAD_SW=true SW=true vite build",
+    "start-destroy": "npm run run-build-destroy && npm run serve",
+    "start-reloadsw-destroy": "npm run run-build-reloadsw-destroy && npm run serve",
+    "start-claims-destroy": "npm run run-build-claims-destroy && npm run serve",
+    "start-claims-reloadsw-destroy": "npm run run-build-reloadsw-claims-destroy && npm run serve",
+    "start-sw-destroy": "npm run run-build-sw-destroy && npm run serve",
+    "start-sw-reloadsw-destroy": "npm run run-build-sw-reloadsw-destroy && npm run serve",
+    "start-sw-claims-destroy": "npm run run-build-sw-claims-destroy && npm run serve",
+    "start-sw-claims-reloadsw-destroy": "npm run run-build-sw-reloadsw-claims-destroy && npm run serve",
+    "run-build-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true vite build",
+    "run-build-reloadsw-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true RELOAD_SW=true vite build",
+    "run-build-claims-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true vite build",
+    "run-build-reloadsw-claims-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true RELOAD_SW=true vite build",
+    "run-build-sw-destroy": "rimraf dist && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true SW=true vite build",
+    "run-build-sw-reloadsw-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true RELOAD_SW=true SW=true vite build",
+    "run-build-sw-claims-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true SW=true vite build",
+    "run-build-sw-reloadsw-claims-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true RELOAD_SW=true SW=true vite build",
     "serve": "serve dist"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router": "^6.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router": "^6.3.0",
     "react-router-config": "^5.1.1",
-    "react-router-dom": "^6.0.2"
+    "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
-    "@rollup/plugin-replace": "^3.0.0",
-    "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-router-config": "^5.0.3",
-    "@types/react-router-dom": "^5.3.2",
+    "@rollup/plugin-replace": "^4.0.0",
+    "@types/react": "^18.0.18",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router-config": "^5.0.6",
+    "@types/react-router-dom": "^5.3.3",
     "@vitejs/plugin-react-refresh": "^1.3.6",
     "cross-env": "^7.0.3",
-    "https-localhost": "^4.7.0",
+    "https-localhost": "^4.7.1",
     "rimraf": "^3.0.2",
-    "typescript": "^4.5.2",
-    "vite": "^2.7.1",
+    "typescript": "^4.8.2",
+    "vite": "^3.1.0",
     "vite-plugin-pwa": "workspace:*",
-    "workbox-core": "^6.4.2",
-    "workbox-precaching": "^6.4.2",
-    "workbox-routing": "^6.4.2",
-    "workbox-window": "^6.4.2"
+    "workbox-core": "^6.5.4",
+    "workbox-precaching": "^6.5.4",
+    "workbox-routing": "^6.5.4",
+    "workbox-window": "^6.5.4"
   }
 }
