@@ -21,28 +24,44 @@
     "run-build-sw": "rimraf dist && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW=true vite build",
     "run-build-sw-reloadsw": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true RELOAD_SW=true vite build",
     "run-build-sw-claims": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true CLAIMS=true SW=true vite build",
-    "run-build-sw-reloadsw-claims": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true CLAIMS=true RELOAD_SW=true SW=true vite build",
+    "run-build-sw-claims-reloadsw": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true CLAIMS=true RELOAD_SW=true SW=true vite build",
+    "start-destroy": "npm run run-build-destroy && npm run serve",
+    "start-reloadsw-destroy": "npm run run-build-reloadsw-destroy && npm run serve",
+    "start-claims-destroy": "npm run run-build-claims-destroy && npm run serve",
+    "start-claims-reloadsw-destroy": "npm run run-build-reloadsw-claims-destroy && npm run serve",
+    "start-sw-destroy": "npm run run-build-sw-destroy && npm run serve",
+    "start-sw-reloadsw-destroy": "npm run run-build-sw-reloadsw-destroy && npm run serve",
+    "start-sw-claims-destroy": "npm run run-build-sw-claims-destroy && npm run serve",
+    "start-sw-claims-reloadsw-destroy": "npm run run-build-sw-reloadsw-claims-destroy && npm run serve",
+    "run-build-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true vite build",
+    "run-build-reloadsw-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true RELOAD_SW=true vite build",
+    "run-build-claims-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true vite build",
+    "run-build-reloadsw-claims-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true RELOAD_SW=true vite build",
+    "run-build-sw-destroy": "rimraf dist && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true SW=true vite build",
+    "run-build-sw-reloadsw-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true RELOAD_SW=true SW=true vite build",
+    "run-build-sw-claims-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true SW=true vite build",
+    "run-build-sw-reloadsw-claims-destroy": "cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true RELOAD_SW=true SW=true vite build",
     "serve": "serve dist"
   },
   "devDependencies": {
-    "@rollup/plugin-replace": "^3.0.0",
-    "@roxi/routify": "^2.18.4",
-    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.31",
-    "@tsconfig/svelte": "^2.0.1",
+    "@rollup/plugin-replace": "^4.0.0",
+    "@roxi/routify": "^2.18.8",
+    "@sveltejs/vite-plugin-svelte": "^1.0.5",
+    "@tsconfig/svelte": "^3.0.0",
     "cross-env": "^7.0.3",
-    "eslint": "^8.4.1",
-    "eslint-plugin-svelte3": "^3.2.1",
-    "https-localhost": "^4.7.0",
+    "eslint": "^8.23.0",
+    "eslint-plugin-svelte3": "^4.0.0",
+    "https-localhost": "^4.7.1",
     "rimraf": "^3.0.2",
-    "svelte": "^3.44.2",
-    "svelte-check": "^2.2.10",
-    "svelte-preprocess": "^4.9.8",
-    "typescript": "^4.5.2",
-    "vite": "^2.7.1",
+    "svelte": "^3.50.0",
+    "svelte-check": "^2.9.0",
+    "svelte-preprocess": "^4.10.7",
+    "typescript": "^4.8.2",
+    "vite": "^3.1.0",
     "vite-plugin-pwa": "workspace:*",
-    "workbox-core": "^6.4.2",
-    "workbox-precaching": "^6.4.2",
-    "workbox-routing": "^6.4.2"
+    "workbox-core": "^6.5.4",
+    "workbox-precaching": "^6.5.4",
+    "workbox-routing": "^6.5.4"
   },
   "eslintConfig": {
     "root": true,
