@@ -25,29 +28,45 @@
     "run-build-sw": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --SW=true",
     "run-build-sw-reloadsw": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true RELOAD_SW=true SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --RELOAD_SW=true --SW=true",
     "run-build-sw-claims": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true CLAIMS=true SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --CLAIMS=true --SW=true",
-    "run-build-sw-reloadsw-claims": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true CLAIMS=true RELOAD_SW=true SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --CLAIMS=true --RELOAD_SW=true --SW=true"
+    "run-build-sw-claims-reloadsw": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true CLAIMS=true RELOAD_SW=true SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --CLAIMS=true --RELOAD_SW=true --SW=true",
+    "start-destroy": "npm run run-build-destroy && npm run preview",
+    "start-reloadsw-destroy": "npm run run-build-reloadsw-destroy && npm run preview",
+    "start-claims-destroy": "npm run run-build-claims-destroy && npm run preview",
+    "start-claims-reloadsw-destroy": "npm run run-build-reloadsw-claims-destroy && npm run preview",
+    "start-sw-destroy": "npm run run-build-sw-destroy && npm run preview",
+    "start-sw-reloadsw-destroy": "npm run run-build-sw-reloadsw-destroy && npm run preview",
+    "start-sw-claims-destroy": "npm run run-build-sw-claims-destroy && npm run preview",
+    "start-sw-claims-reloadsw-destroy": "npm run run-build-sw-reloadsw-claims-destroy && npm run preview",
+    "run-build-destroy": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --SW_DESTROY=true",
+    "run-build-reloadsw-destroy": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true RELOAD_SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --SW_DESTROY=true --RELOAD_SW=true",
+    "run-build-claims-destroy": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --SW_DESTROY=true --CLAIMS=true",
+    "run-build-reloadsw-claims-destroy": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true RELOAD_SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --SW_DESTROY=true --CLAIMS=true --RELOAD_SW=true",
+    "run-build-sw-destroy": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --SW_DESTROY=true --SW=true",
+    "run-build-sw-reloadsw-destroy": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true RELOAD_SW=true SW=true svelte-kit build && node ./pwa.jss --BASE_URL=/ --SOURCE_MAP=true --SW_DESTROY=true --RELOAD_SW=true --SW=true",
+    "run-build-sw-claims-destroy": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --SW_DESTROY=true --CLAIMS=true --SW=true",
+    "run-build-sw-reloadsw-claims-destroy": "rimraf .svelte-kit && rimraf build && cross-env DEBUG=vite-plugin-pwa:* BASE_URL=/ SOURCE_MAP=true SW_DESTROY=true CLAIMS=true RELOAD_SW=true SW=true svelte-kit build && node ./pwa.js --BASE_URL=/ --SOURCE_MAP=true --SW_DESTROY=true --CLAIMS=true --RELOAD_SW=true --SW=true"
   },
   "devDependencies": {
-    "@rollup/plugin-replace": "^3.0.0",
+    "@rollup/plugin-replace": "^4.0.0",
     "@sveltejs/adapter-static": "next",
     "@sveltejs/kit": "next",
-    "@typescript-eslint/eslint-plugin": "^5.6.0",
-    "@typescript-eslint/parser": "^5.6.0",
+    "@typescript-eslint/eslint-plugin": "^5.30.3",
+    "@typescript-eslint/parser": "^5.30.3",
     "cross-env": "^7.0.3",
-    "eslint": "^8.4.1",
-    "eslint-plugin-svelte3": "^3.2.1",
-    "https-localhost": "^4.7.0",
-    "minimist": "^1.2.5",
+    "eslint": "^8.19.0",
+    "eslint-plugin-svelte3": "^4.0.0",
+    "https-localhost": "^4.7.1",
+    "minimist": "^1.2.6",
     "rimraf": "^3.0.2",
-    "svelte": "^3.44.2",
-    "svelte-check": "^2.2.10",
-    "svelte-preprocess": "^4.9.8",
-    "tslib": "^2.3.1",
-    "typescript": "^4.5.2",
+    "svelte": "^3.48.0",
+    "svelte-check": "^2.8.0",
+    "svelte-preprocess": "^4.10.7",
+    "tslib": "^2.4.0",
+    "typescript": "^4.7.4",
     "vite-plugin-pwa": "workspace:*",
-    "workbox-core": "^6.4.2",
-    "workbox-precaching": "^6.4.2",
-    "workbox-routing": "^6.4.2"
+    "workbox-core": "^6.5.3",
+    "workbox-precaching": "^6.5.3",
+    "workbox-routing": "^6.5.3"
   },
   "type": "module"
 }
