@@ -10,17 +10,16 @@
     "serve": "serve dist"
   },
   "dependencies": {
-    "vue": "^3.2.24"
+    "vue": "^3.2.38"
   },
   "devDependencies": {
-    "@rollup/plugin-replace": "^3.0.0",
-    "@vitejs/plugin-vue": "^1.10.2",
-    "@vue/compiler-sfc": "^3.2.24",
-    "@vueuse/core": "^7.2.2",
+    "@rollup/plugin-replace": "^4.0.0",
+    "@vitejs/plugin-vue": "^3.1.0",
+    "@vueuse/core": "^9.2.0",
     "cross-env": "^7.0.3",
-    "https-localhost": "^4.7.0",
-    "typescript": "^4.5.2",
-    "vite": "^2.7.1",
+    "https-localhost": "^4.7.1",
+    "typescript": "^4.8.2",
+    "vite": "^3.1.0",
     "vite-plugin-pwa": "workspace:*"
   }
 }
