@@ -1,12 +1,17 @@
 {
   "name": "vite-plugin-pwa",
-  "version": "0.11.12",
+  "version": "0.13.1",
+  "packageManager": "pnpm@7.12.0",
   "description": "Zero-config PWA for Vite",
+  "author": "antfu <anthonyfu117@hotmail.com>",
   "license": "MIT",
+  "funding": "https://github.com/sponsors/antfu",
+  "homepage": "https://github.com/antfu/vite-plugin-pwa",
   "repository": {
     "type": "git",
     "url": "https://github.com/antfu/vite-plugin-pwa"
   },
+  "bugs": "https://github.com/antfu/vite-plugin-pwa/issues",
   "keywords": [
     "react",
     "pwa",
@@ -20,22 +25,36 @@
     "vite",
     "vite-plugin"
   ],
-  "funding": "https://github.com/sponsors/antfu",
-  "author": "antfu <anthonyfu117@hotmail.com>",
-  "files": [
-    "dist",
-    "index.d.ts",
-    "client.d.ts"
-  ],
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.js",
+      "import": "./dist/index.mjs"
+    },
+    "./*": "./*"
+  },
   "main": "dist/index.js",
   "module": "dist/index.mjs",
   "types": "index.d.ts",
-  "homepage": "https://github.com/antfu/vite-plugin-pwa",
-  "bugs": "https://github.com/antfu/vite-plugin-pwa/issues",
+  "typesVersions": {
+    "*": {
+      "*": [
+        "./dist/*",
+        "./*"
+      ]
+    }
+  },
+  "files": [
+    "dist",
+    "*.d.ts"
+  ],
   "scripts": {
     "docs": "npm -C docs run dev",
     "docs:build": "npm -C docs run build",
     "docs:serve": "npm -C docs run serve",
+    "docs:preview": "npm -C docs run preview",
+    "lint": "eslint .",
+    "lint-fix": "nr lint --fix",
     "dev": "esno scripts/dev.ts",
     "build": "esno scripts/build.ts",
     "prepublishOnly": "npm run build",
@@ -46,38 +65,39 @@
     "example:vue:start:cdn": "npm -C examples/vue-basic-cdn run start"
   },
   "peerDependencies": {
-    "vite": "^2.0.0",
-    "workbox-build": "^6.4.0",
-    "workbox-window": "^6.4.0"
+    "vite": "^3.1.0",
+    "workbox-build": "^6.5.4",
+    "workbox-window": "^6.5.4"
   },
   "dependencies": {
-    "debug": "^4.3.3",
-    "fast-glob": "^3.2.7",
-    "pretty-bytes": "^5.6.0",
-    "rollup": "^2.60.2",
-    "workbox-build": "^6.4.2",
-    "workbox-window": "^6.4.2"
+    "debug": "^4.3.4",
+    "fast-glob": "^3.2.11",
+    "pretty-bytes": "^6.0.0",
+    "rollup": "^2.79.0",
+    "workbox-build": "^6.5.4",
+    "workbox-window": "^6.5.4"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^0.12.2",
-    "@antfu/ni": "^0.11.1",
+    "@antfu/eslint-config": "^0.26.3",
+    "@antfu/ni": "^0.18.0",
     "@types/debug": "^4.1.7",
+    "@types/node": "^18.7.15",
     "@types/prompts": "^2.4.0",
-    "@types/react": "^17.0.37",
+    "@types/react": "^18.0.18",
     "@types/workbox-build": "^5.0.1",
-    "@typescript-eslint/eslint-plugin": "^5.6.0",
-    "eslint": "^8.4.1",
-    "esno": "^0.12.1",
-    "kolorist": "^1.5.0",
-    "preact": "^10.6.3",
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "eslint": "^8.23.0",
+    "esno": "0.16.3",
+    "kolorist": "^1.5.1",
+    "preact": "^10.10.6",
     "prompts": "^2.4.2",
-    "react": "^17.0.2",
-    "rollup": "^2.60.2",
-    "solid-js": "^1.2.6",
-    "svelte": "^3.44.2",
-    "tsup": "^5.10.3",
-    "typescript": "^4.5.2",
-    "vite": "^2.7.1",
-    "vue": "^3.2.24"
+    "react": "^18.2.0",
+    "rollup": "^2.79.0",
+    "solid-js": "^1.5.4",
+    "svelte": "^3.50.0",
+    "tsup": "^6.2.3",
+    "typescript": "^4.8.2",
+    "vite": "^3.1.0",
+    "vue": "^3.2.38"
   }
 }
