@@ -1,14 +1,13 @@
-// eslint-disable-next-line no-use-before-define
 import React from 'react'
-import ReactDOM from 'react-dom'
-import { BrowserRouter, Routes, Route } from 'react-router-dom'
+import { BrowserRouter, Route, Routes } from 'react-router-dom'
+import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App'
 import Home from './pages/Home'
 import About from './pages/About'
 import Hi from './pages/hi/[name]'
 
-ReactDOM.render(
+createRoot(document.getElementById('app')!).render(
   <BrowserRouter>
     <Routes>
       <Route path="/" element={<App />}>
@@ -20,5 +19,4 @@ ReactDOM.render(
       </Route>
     </Routes>
   </BrowserRouter>,
-  document.getElementById('app'),
 )
