@@ -1,9 +1,22 @@
 declare module 'virtual:pwa-register' {
-  export type RegisterSWOptions = {
+  export interface RegisterSWOptions {
     immediate?: boolean
     onNeedRefresh?: () => void
     onOfflineReady?: () => void
+    /**
+     * Called only if `onRegisteredSW` is not provided.
+     *
+     * @deprecated Use `onRegisteredSW` instead.
+     * @param registration The service worker registration if available.
+     */
     onRegistered?: (registration: ServiceWorkerRegistration | undefined) => void
+    /**
+     * Called once the service worker is registered (requires version `0.12.8+`).
+     *
+     * @param swScriptUrl The service worker script url.
+     * @param registration The service worker registration if available.
+     */
+    onRegisteredSW?: (swScriptUrl: string, registration: ServiceWorkerRegistration | undefined) => void
     onRegisterError?: (error: any) => void
   }
 
@@ -11,14 +24,28 @@ declare module 'virtual:pwa-register' {
 }
 
 declare module 'virtual:pwa-register/vue' {
+  // eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error
   // @ts-ignore ignore when vue is not installed
   import type { Ref } from 'vue'
 
-  export type RegisterSWOptions = {
+  export interface RegisterSWOptions {
     immediate?: boolean
     onNeedRefresh?: () => void
     onOfflineReady?: () => void
+    /**
+     * Called only if `onRegisteredSW` is not provided.
+     *
+     * @deprecated Use `onRegisteredSW` instead.
+     * @param registration The service worker registration if available.
+     */
     onRegistered?: (registration: ServiceWorkerRegistration | undefined) => void
+    /**
+     * Called once the service worker is registered (requires version `0.12.8+`).
+     *
+     * @param swScriptUrl The service worker script url.
+     * @param registration The service worker registration if available.
+     */
+    onRegisteredSW?: (swScriptUrl: string, registration: ServiceWorkerRegistration | undefined) => void
     onRegisterError?: (error: any) => void
   }
 
@@ -30,14 +57,28 @@ declare module 'virtual:pwa-register/vue' {
 }
 
 declare module 'virtual:pwa-register/svelte' {
+  // eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error
   // @ts-ignore ignore when svelte is not installed
-  import { Writable } from 'svelte/store'
+  import type { Writable } from 'svelte/store'
 
-  export type RegisterSWOptions = {
+  export interface RegisterSWOptions {
     immediate?: boolean
     onNeedRefresh?: () => void
     onOfflineReady?: () => void
+    /**
+     * Called only if `onRegisteredSW` is not provided.
+     *
+     * @deprecated Use `onRegisteredSW` instead.
+     * @param registration The service worker registration if available.
+     */
     onRegistered?: (registration: ServiceWorkerRegistration | undefined) => void
+    /**
+     * Called once the service worker is registered (requires version `0.12.8+`).
+     *
+     * @param swScriptUrl The service worker script url.
+     * @param registration The service worker registration if available.
+     */
+    onRegisteredSW?: (swScriptUrl: string, registration: ServiceWorkerRegistration | undefined) => void
     onRegisterError?: (error: any) => void
   }
 
@@ -49,14 +90,28 @@ declare module 'virtual:pwa-register/svelte' {
 }
 
 declare module 'virtual:pwa-register/react' {
+  // eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error
   // @ts-ignore ignore when react is not installed
-  import { Dispatch, SetStateAction } from 'react'
+  import type { Dispatch, SetStateAction } from 'react'
 
-  export type RegisterSWOptions = {
+  export interface RegisterSWOptions {
     immediate?: boolean
     onNeedRefresh?: () => void
     onOfflineReady?: () => void
+    /**
+     * Called only if `onRegisteredSW` is not provided.
+     *
+     * @deprecated Use `onRegisteredSW` instead.
+     * @param registration The service worker registration if available.
+     */
     onRegistered?: (registration: ServiceWorkerRegistration | undefined) => void
+    /**
+     * Called once the service worker is registered (requires version `0.12.8+`).
+     *
+     * @param swScriptUrl The service worker script url.
+     * @param registration The service worker registration if available.
+     */
+    onRegisteredSW?: (swScriptUrl: string, registration: ServiceWorkerRegistration | undefined) => void
     onRegisterError?: (error: any) => void
   }
 
@@ -68,14 +123,28 @@ declare module 'virtual:pwa-register/react' {
 }
 
 declare module 'virtual:pwa-register/solid' {
+  // eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error
   // @ts-ignore ignore when solid-js is not installed
-  import { Accessor, Setter } from 'solid-js'
+  import type { Accessor, Setter } from 'solid-js'
 
-  export type RegisterSWOptions = {
+  export interface RegisterSWOptions {
     immediate?: boolean
     onNeedRefresh?: () => void
     onOfflineReady?: () => void
+    /**
+     * Called only if `onRegisteredSW` is not provided.
+     *
+     * @deprecated Use `onRegisteredSW` instead.
+     * @param registration The service worker registration if available.
+     */
     onRegistered?: (registration: ServiceWorkerRegistration | undefined) => void
+    /**
+     * Called once the service worker is registered (requires version `0.12.8+`).
+     *
+     * @param swScriptUrl The service worker script url.
+     * @param registration The service worker registration if available.
+     */
+    onRegisteredSW?: (swScriptUrl: string, registration: ServiceWorkerRegistration | undefined) => void
     onRegisterError?: (error: any) => void
   }
 
@@ -87,14 +156,28 @@ declare module 'virtual:pwa-register/solid' {
 }
 
 declare module 'virtual:pwa-register/preact' {
+  // eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error
   // @ts-ignore ignore when preact/hooks is not installed
-  import { StateUpdater } from 'preact/hooks'
+  import type { StateUpdater } from 'preact/hooks'
 
-  export type RegisterSWOptions = {
+  export interface RegisterSWOptions {
     immediate?: boolean
     onNeedRefresh?: () => void
     onOfflineReady?: () => void
+    /**
+     * Called only if `onRegisteredSW` is not provided.
+     *
+     * @deprecated Use `onRegisteredSW` instead.
+     * @param registration The service worker registration if available.
+     */
     onRegistered?: (registration: ServiceWorkerRegistration | undefined) => void
+    /**
+     * Called once the service worker is registered (requires version `0.12.8+`).
+     *
+     * @param swScriptUrl The service worker script url.
+     * @param registration The service worker registration if available.
+     */
+    onRegisteredSW?: (swScriptUrl: string, registration: ServiceWorkerRegistration | undefined) => void
     onRegisterError?: (error: any) => void
   }
 
