@@ -1,15 +1,14 @@
 <template>
   <p class="prompt-img">
-    <img src="/prompt-update.png" alt="Prompt for update for new content dialog image with the update and cancel button" width="720" height="178">
+    <img src="/prompt-update.png" alt="Prompt for update for new content dialog image with the update and cancel button" width="922" height="228">
   </p>
-
 </template>
 
-<style scoped lang="postcss">
+<style scoped>
 .prompt-img {
   text-align: center;
-  img {
-    object-fit: scale-down;
-  }
+}
+.prompt-img img {
+  object-fit: scale-down;
 }
 </style>
