@@ -1,33 +1,51 @@
 <script setup lang="ts">
-import { useRegisterSW } from 'virtual:pwa-register/vue'
+import { onBeforeMount, ref } from 'vue'
 
-const { needRefresh, updateServiceWorker } = useRegisterSW()
+const needRefresh = ref(false)
 
-const close = async() => {
+let updateServiceWorker: (() => Promise<void>) | undefined
+
+const onNeedRefresh = () => {
+  needRefresh.value = true
+}
+const close = async () => {
   needRefresh.value = false
 }
+
+onBeforeMount(async () => {
+  const { registerSW } = await import('virtual:pwa-register')
+  updateServiceWorker = registerSW({
+    immediate: true,
+    onNeedRefresh,
+  })
+})
 </script>
 
 <template>
-  <div
-    v-if="needRefresh"
-    class="pwa-toast bg-$c-bg border border-$c-divider fixed right-0 bottom-0 m-6 px-6 py-4 rounded shadow-xl"
-    role="alert"
-  >
-    <div class="mb-3">
-      New content available, click the reload button to update.
+  <Teleport to="body">
+    <div
+      v-if="needRefresh"
+      class="pwa-toast z-100 bg-$vp-c-bg border border-$pwa-divider fixed right-0 bottom-0 m-6 px-6 py-4 rounded shadow-xl"
+      role="alertdialog"
+      aria-labelledby="pwa-message"
+    >
+      <div id="pwa-message" class="mb-3">
+        New content available, click the reload button to update.
+      </div>
+      <button
+        type="button"
+        class="pwa-refresh mr-2 px-3 py-1 rounded"
+        @click="updateServiceWorker?.()"
+      >
+        Reload
+      </button>
+      <button
+        type="button"
+        class="pwa-cancel border border-$pwa-divider mr-2 px-3 py-1 rounded"
+        @click="close"
+      >
+        Close
+      </button>
     </div>
-    <button
-      class="border border-$c-divider bg-$c-brand text-black mr-2 px-3 py-1 rounded hover:bg-$c-brand-light"
-      @click="updateServiceWorker()"
-    >
-      Reload
-    </button>
-    <button
-      class="border border-$c-divider bg-$c-brand text-black mr-2 px-3 py-1 rounded hover:bg-$c-brand-light"
-      @click="close"
-    >
-      Close
-    </button>
-  </div>
+  </Teleport>
 </template>
