@@ -14,10 +14,10 @@ const preconnect = `
     <link rel="preconnect" crossorigin="anonymous" href="${gstatic}">
 `
 
-export const optimizePages = async() => {
+export const optimizePages = async () => {
   const names = await fg('./.vitepress/dist/**/*.html', { onlyFiles: true })
 
-  await Promise.all(names.map(async(i) => {
+  await Promise.all(names.map(async (i) => {
     let html = await fs.readFile(i, 'utf-8')
 
     let preloadImg = '\n\t<link rel="prefetch" href="/icon_light.svg">\n\t<link rel="prefetch" href="/icon_dark.svg">'
@@ -43,20 +43,21 @@ export const optimizePages = async() => {
     <noscript>
       <link rel="stylesheet" crossorigin="anonymous" href="${firaFont}" />
     </noscript>
-    <link rel="prefetch" href="/manifest.webmanifest">${preloadImg}\n`).trim()
+    <link rel="prefetch" href="/manifest.webmanifest">${preloadImg}
+    <link rel="manifest" href="/manifest.webmanifest">\n`).trim()
 
-    html = html.replace(
-      '</head>',
-      '\t<link rel="manifest" href="/manifest.webmanifest">\n<script>\n'
-        + '    (function() {\n'
-        + '      const prefersDark = window.matchMedia && window.matchMedia(\'(prefers-color-scheme: dark)\').matches\n'
-        + '      const setting = localStorage.getItem(\'color-schema\') || \'auto\'\n'
-        + '      if (setting === \'dark\' || (prefersDark && setting !== \'light\'))\n'
-        + '        document.documentElement.classList.toggle(\'dark\', true)\n'
-        + '    })()\n'
-        + '  </script></head>',
-    )
-
+    // html = html.replace(
+    //   '</head>',
+    //   '\t<link rel="manifest" href="/manifest.webmanifest">\n<script>\n'
+    //     + '    (function() {\n'
+    //     + '      const prefersDark = window.matchMedia && window.matchMedia(\'(prefers-color-scheme: dark)\').matches\n'
+    //     + '      const setting = localStorage.getItem(\'vueuse-color-schema\') || \'auto\'\n'
+    //     + '      if (setting === \'dark\' || (prefersDark && setting !== \'light\'))\n'
+    //     + '        document.documentElement.classList.toggle(\'dark\', true)\n'
+    //     + '    })()\n'
+    //     + '  </script></head>',
+    // )
+    //
     html = html.replace(/aria-hidden="true"/gi, 'tabindex="-1" aria-hidden="true"')
 
     await fs.writeFile(i, html, 'utf-8')
