@@ -1,15 +1,13 @@
 import { resolveConfig } from 'vite'
-import { VitePluginPWAAPI } from '../../dist'
+import type { VitePluginPWAAPI } from 'vite-plugin-pwa'
 import { optimizePages } from './assets'
+import { pwa } from './pwa'
 
-const rebuildPwa = async() => {
-  const config = await resolveConfig({}, 'build', 'production')
+export const buildEnd = async () => {
+  await optimizePages()
+  const config = await resolveConfig({ plugins: [pwa()] }, 'build', 'production')
   // when `vite-plugin-pwa` is presented, use it to regenerate SW after rendering
   const pwaPlugin: VitePluginPWAAPI = config.plugins.find(i => i.name === 'vite-plugin-pwa')?.api
-  if (pwaPlugin && pwaPlugin.generateSW && !pwaPlugin.disabled) {
-    await optimizePages()
+  if (pwaPlugin && pwaPlugin.generateSW && !pwaPlugin.disabled)
     await pwaPlugin.generateSW()
-  }
 }
-
-rebuildPwa()
