@@ -1,22 +1,26 @@
 import { defineConfig } from 'vite'
 import Components from 'unplugin-vue-components/vite'
-import WindiCSS from 'vite-plugin-windicss'
-import Icons from 'unplugin-icons/vite'
-import IconsResolver from 'unplugin-icons/resolver'
-import replace from '@rollup/plugin-replace'
-import { VitePWA } from '../dist'
-import { version } from '../package.json'
+import { presetAttributify, presetUno } from 'unocss'
+import Unocss from 'unocss/vite'
+import NavbarFix from './plugins/navbar'
 
 export default defineConfig({
+  ssr: {
+    format: 'cjs',
+  },
+  legacy: {
+    buildSsrCjsExternalHeuristics: true,
+  },
   build: {
+    // sourcemap: true,
+    // minify: false,
     ssrManifest: false,
     manifest: false,
   },
   optimizeDeps: {
     exclude: [
-      'vue',
-      'vue-global-api',
       '@vueuse/core',
+      'vitepress',
     ],
   },
   server: {
@@ -25,11 +29,6 @@ export default defineConfig({
     },
   },
   plugins: [
-    replace({
-      preventAssignment: true,
-      __PWA_VERSION__: version,
-    }),
-
     // https://github.com/antfu/vite-plugin-components
     Components({
       dirs: [
@@ -42,95 +41,14 @@ export default defineConfig({
       include: [/\.vue$/, /\.vue\?vue/, /\.md$/],
 
       // generate `components.d.ts` for ts support with Volar
-      dts: false,
-      // auto import icons
-      resolvers: [
-        // https://github.com/antfu/vite-plugin-icons
-        IconsResolver({
-          componentPrefix: '',
-          // enabledCollections: ['carbon'],
-        }),
-      ],
+      dts: '.vitepress/components.d.ts',
     }),
 
-    // https://github.com/antfu/vite-plugin-icons
-    Icons(),
+    NavbarFix(),
 
-    // https://github.com/antfu/vite-plugin-windicss
-    WindiCSS(),
-
-    // https://github.com/antfu/vite-plugin-pwa
-    VitePWA({
-      outDir: '.vitepress/dist',
-      registerType: 'prompt',
-      includeAssets: [
-        'favicon.svg',
-        'favicon.ico',
-        'robots.txt',
-        'safari-pinned-tab.svg',
-        'banner_light.svg',
-        'banner_dark.svg',
-        'icon_light.svg',
-        'icon_dark.svg',
-        'prompt-update.png',
-      ],
-      manifest: {
-        id: '/',
-        name: 'Vite Plugin PWA',
-        short_name: 'PWA for Vite',
-        description: 'Zero-config PWA for Vite',
-        theme_color: '#ffffff',
-        icons: [
-          {
-            src: 'pwa-192x192.png',
-            sizes: '192x192',
-            type: 'image/png',
-          },
-          {
-            src: 'pwa-512x512.png',
-            sizes: '512x512',
-            type: 'image/png',
-          },
-          {
-            src: 'icon_light.svg',
-            sizes: '155x155',
-            type: 'image/svg',
-            purpose: 'any maskable',
-          },
-        ],
-      },
-      workbox: {
-        runtimeCaching: [
-          {
-            urlPattern: /^https:\/\/fonts\.googleapis\.com\/.*/i,
-            handler: 'CacheFirst',
-            options: {
-              cacheName: 'google-fonts-cache',
-              expiration: {
-                maxEntries: 10,
-                maxAgeSeconds: 60 * 60 * 24 * 365, // <== 365 days
-              },
-              cacheableResponse: {
-                statuses: [0, 200],
-              },
-            },
-          },
-          {
-            urlPattern: /^https:\/\/fonts\.gstatic\.com\/.*/i,
-            handler: 'CacheFirst',
-            options: {
-              cacheName: 'gstatic-fonts-cache',
-              expiration: {
-                maxEntries: 10,
-                maxAgeSeconds: 60 * 60 * 24 * 365, // <== 365 days
-              },
-              cacheableResponse: {
-                statuses: [0, 200],
-              },
-            },
-          },
-        ],
-      },
+    // https://github.com/unocss/unocss
+    Unocss({
+      presets: [presetUno(), presetAttributify()],
     }),
   ],
 })
