@@ -1,4 +1,3 @@
-// eslint-disable-next-line no-use-before-define
 import { useState } from 'preact/hooks'
 import { route } from 'preact-router'
 import './Home.css'
@@ -8,12 +7,12 @@ function Home() {
 
   const [name, setName] = useState('')
 
-  // @ts-ignore
+  // @ts-expect-error just ignore
   const handleChange = (event) => {
     setName(event.target.value || '')
   }
 
-  // @ts-ignore
+  // @ts-expect-error just ignore
   const handleSubmit = (event) => {
     event.preventDefault()
     if (name)
