@@ -1,8 +1,11 @@
-// eslint-disable-next-line no-use-before-define
 import React from 'react'
 import './ReloadPrompt.css'
 
 import { useRegisterSW } from 'virtual:pwa-register/react'
+import { pwaInfo } from 'virtual:pwa-info'
+
+// eslint-disable-next-line no-console
+console.log(pwaInfo)
 
 function ReloadPrompt() {
   // replaced dynamically
@@ -15,20 +18,24 @@ function ReloadPrompt() {
     needRefresh: [needRefresh, setNeedRefresh],
     updateServiceWorker,
   } = useRegisterSW({
-    onRegistered(r) {
-      // @ts-ignore
+    onRegisteredSW(swUrl, r) {
+      // eslint-disable-next-line no-console
+      console.log(`Service Worker at: ${swUrl}`)
+      // @ts-expect-error just ignore
       if (reloadSW === 'true') {
         r && setInterval(() => {
+          // eslint-disable-next-line no-console
           console.log('Checking for sw update')
           r.update()
         }, 20000 /* 20s for testing purposes */)
       }
       else {
-        // eslint-disable-next-line prefer-template
+        // eslint-disable-next-line prefer-template,no-console
         console.log('SW Registered: ' + r)
       }
     },
     onRegisterError(error) {
+      // eslint-disable-next-line no-console
       console.log('SW registration error', error)
     },
   })
