@@ -1,4 +1,3 @@
-// eslint-disable-next-line no-use-before-define
 import React, { useState } from 'react'
 import { useNavigate } from 'react-router'
 import './Home.css'
@@ -10,12 +9,12 @@ function Home() {
 
   const router = useNavigate()
 
-  // @ts-ignore
+  // @ts-expect-error just ignore
   const handleChange = (event) => {
     setName(event.target.value || '')
   }
 
-  // @ts-ignore
+  // @ts-expect-error just ignore
   const handleSubmit = (event) => {
     event.preventDefault()
     if (name)
