@@ -1,6 +1,7 @@
 import { defineConfig } from 'vite'
 import reactRefresh from '@vitejs/plugin-react-refresh'
-import { ManifestOptions, VitePWA, VitePWAOptions } from 'vite-plugin-pwa'
+import type { ManifestOptions, VitePWAOptions } from 'vite-plugin-pwa'
+import { VitePWA } from 'vite-plugin-pwa'
 import replace from '@rollup/plugin-replace'
 
 const pwaOptions: Partial<VitePWAOptions> = {
@@ -30,11 +31,18 @@ const pwaOptions: Partial<VitePWAOptions> = {
       },
     ],
   },
+  devOptions: {
+    enabled: process.env.SW_DEV === 'true',
+    /* when using generateSW the PWA plugin will switch to classic */
+    type: 'module',
+    navigateFallback: 'index.html',
+  },
 }
 
 const replaceOptions = { __DATE__: new Date().toISOString() }
 const claims = process.env.CLAIMS === 'true'
 const reload = process.env.RELOAD_SW === 'true'
+const selfDestroying = process.env.SW_DESTROY === 'true'
 
 if (process.env.SW === 'true') {
   pwaOptions.srcDir = 'src'
@@ -48,9 +56,12 @@ if (claims)
   pwaOptions.registerType = 'autoUpdate'
 
 if (reload) {
-  // @ts-ignore
+  // @ts-expect-error just ignore
   replaceOptions.__RELOAD_SW__ = 'true'
 }
+
+if (selfDestroying)
+  pwaOptions.selfDestroying = selfDestroying
 
 export default defineConfig({
   // base: process.env.BASE_URL || 'https://github.com/',
