@@ -1,4 +1,4 @@
-/* eslint-disable react/react-in-jsx-scope,react/no-unknown-property */
+/* eslint-disable react/no-unknown-property */
 import type { Component } from 'solid-js'
 import { useRoutes } from 'solid-app-router'
 import { routes } from './routes'
