@@ -1,4 +1,4 @@
-/* eslint-disable react/react-in-jsx-scope,react/no-unknown-property */
+/* eslint-disable react/no-unknown-property */
 import { createSignal } from 'solid-js'
 import { useNavigate } from 'solid-app-router'
 import styles from './home.module.css'
