@@ -13,8 +13,10 @@ const config = {
 	kit: {
 		adapter: adapter(),
 
-		// hydrate the <div id="svelte"> element in src/app.html
-		target: '#svelte',
+		prerender: {
+			default: true,
+		},
+
 		vite: {
 			plugins: [
 				VitePWA(pwaConfiguration),
