@@ -1,5 +1,5 @@
 declare module '*.vue' {
-  import { defineComponent } from 'vue'
+  import type { defineComponent } from 'vue'
   const Component: ReturnType<typeof defineComponent>
   export default Component
 }
