@@ -1,6 +1,9 @@
-
 <script setup lang="ts">
 import { useRegisterSW } from 'virtual:pwa-register/vue'
+import { pwaInfo } from 'virtual:pwa-info'
+
+// eslint-disable-next-line no-console
+console.log(pwaInfo)
 
 // replaced dyanmicaly
 const reloadSW: any = '__RELOAD_SW__'
@@ -11,24 +14,27 @@ const {
   updateServiceWorker,
 } = useRegisterSW({
   immediate: true,
-  onRegistered(r) {
+  onRegisteredSW(swUrl, r) {
+    // eslint-disable-next-line no-console
+    console.log(`Service Worker at: ${swUrl}`)
     if (reloadSW === 'true') {
-      r && setInterval(async() => {
+      r && setInterval(async () => {
+        // eslint-disable-next-line no-console
         console.log('Checking for sw update')
         await r.update()
       }, 20000 /* 20s for testing purposes */)
     }
     else {
+      // eslint-disable-next-line no-console
       console.log(`SW Registered: ${r}`)
     }
   },
 })
 
-const close = async() => {
+const close = async () => {
   offlineReady.value = false
   needRefresh.value = false
 }
-
 </script>
 
 <template>
