@@ -1,5 +1,5 @@
 import { createApp, defineAsyncComponent } from 'vue'
-import { createWebHistory, createRouter } from 'vue-router'
+import { createRouter, createWebHistory } from 'vue-router'
 import App from './App.vue'
 import './index.css'
 
