@@ -1,13 +1,13 @@
 import { useState } from 'react'
-import { RegisterSWOptions } from '../type'
+import type { RegisterSWOptions } from '../type'
 
 export type { RegisterSWOptions }
 
-export function useRegisterSW(options: RegisterSWOptions = {}) {
+export function useRegisterSW(_options: RegisterSWOptions = {}) {
   const needRefresh = useState(false)
   const offlineReady = useState(false)
 
-  const updateServiceWorker = (reloadPage?: boolean) => {}
+  const updateServiceWorker = (_reloadPage?: boolean) => {}
 
   return {
     needRefresh,
