@@ -1,7 +1,7 @@
-import { RegisterSWOptions } from '../type'
+import type { RegisterSWOptions } from '../type'
 
 export type { RegisterSWOptions }
 
-export function registerSW(options: RegisterSWOptions = {}) {
-  return (reloadPage = true) => {}
+export function registerSW(_options: RegisterSWOptions = {}) {
+  return (_reloadPage = true) => {}
 }
