@@ -1,13 +1,13 @@
 import { ref } from 'vue'
-import { RegisterSWOptions } from '../type'
+import type { RegisterSWOptions } from '../type'
 
 export type { RegisterSWOptions }
 
-export function useRegisterSW(options: RegisterSWOptions = {}) {
+export function useRegisterSW(_options: RegisterSWOptions = {}) {
   const needRefresh = ref(false)
   const offlineReady = ref(false)
 
-  const updateServiceWorker = (reloadPage?: boolean) => {}
+  const updateServiceWorker = (_reloadPage?: boolean) => {}
 
   return {
     updateServiceWorker,
