@@ -1,6 +1,30 @@
-/* eslint-disable no-use-before-define */
+import type { ResolvedConfig } from 'vite'
 import type { GenerateSWOptions, InjectManifestOptions, ManifestEntry } from 'workbox-build'
 import type { OutputBundle } from 'rollup'
+
+export type InjectManifestVitePlugins = string[] | ((vitePluginIds: string[]) => string[])
+export type CustomInjectManifestOptions = InjectManifestOptions & {
+  /**
+   * Configure the format to use in the Rollup build.
+   *
+   * @default 'es'
+   */
+  rollupFormat?: 'es' | 'iife'
+  /**
+   * `Vite` plugin ids to use on `Rollup` build.
+   *
+   * **WARN**: this option is for advanced usage, beware, you can break the service worker build.
+   */
+  vitePlugins?: InjectManifestVitePlugins
+}
+
+export interface PWAIntegration {
+  closeBundleOrder?: 'pre' | 'post' | null
+  configureOptions?: (
+    viteOptions: ResolvedConfig,
+    options: Partial<VitePWAOptions>,
+  ) => void | Promise<void>
+}
 
 /**
  * Plugin options.
@@ -24,6 +48,10 @@ export interface VitePWAOptions {
    * @default 'sw.js'
    */
   filename?: string
+  /**
+   * @default 'manifest.webmanifest'
+   */
+  manifestFilename?: string
   /**
    * @default 'generateSW'
    */
@@ -84,7 +112,7 @@ export interface VitePWAOptions {
   /**
    * The workbox object for `injectManifest`
    */
-  injectManifest: Partial<InjectManifestOptions>
+  injectManifest: Partial<CustomInjectManifestOptions>
   /**
    * Override Vite's base options only for PWA
    *
@@ -113,6 +141,20 @@ export interface VitePWAOptions {
    * @default false
    */
   disable: boolean
+  /**
+   * Vite PWA Integration.
+   */
+  integration?: PWAIntegration
+  /**
+   * Development options.
+   */
+  devOptions?: DevOptions
+  /**
+   * Unregister the service worker?
+   *
+   * @default false
+   */
+  selfDestroying?: boolean
 }
 
 export interface ResolvedVitePWAOptions extends Required<VitePWAOptions> {
@@ -120,6 +162,8 @@ export interface ResolvedVitePWAOptions extends Required<VitePWAOptions> {
   swDest: string
   workbox: GenerateSWOptions
   injectManifest: InjectManifestOptions
+  rollupFormat: 'es' | 'iife'
+  vitePlugins: InjectManifestVitePlugins
 }
 
 export interface ManifestOptions {
@@ -140,6 +184,10 @@ export interface ManifestOptions {
    */
   icons: Record<string, any>[]
   /**
+   *
+   */
+  file_handlers: Record<string, any>[]
+  /**
    * @default `routerBase + '?standalone=true'`
    */
   start_url: string
@@ -159,6 +207,10 @@ export interface ManifestOptions {
    * @default `standalone`
    */
   display: string
+  /**
+   * @default []
+   */
+  display_override: string[]
   /**
    * @default `#ffffff`
    */
@@ -206,7 +258,7 @@ export interface ManifestOptions {
     short_name?: string
     url: string
     description?: string
-    icons: string[]
+    icons: Record<string, any>[]
   }[]
   /**
    * @default []
@@ -228,16 +280,71 @@ export interface ManifestOptions {
   iarc_rating_id: string
 }
 
+export interface WebManifestData {
+  href: string
+  useCredentials: boolean
+  /**
+   * Returns the corresponding link tag: `<link rel="manifest" href="<webManifestUrl>" />`.
+   */
+  toLinkTag: () => string
+}
+
+export interface RegisterSWData {
+  shouldRegisterSW: boolean
+  /**
+   * When this flag is `true` the service worker must be registered via inline script otherwise registered via script with src attribute `registerSW.js` .
+   */
+  inline: boolean
+  /**
+   * The path for the inline script: will contain the service worker url.
+   */
+  inlinePath: string
+  /**
+   * The path for the src script for `registerSW.js`.
+   */
+  registerPath: string
+  /**
+   * The scope for the service worker: only required for `inline: true`.
+   */
+  scope: string
+  /**
+   * The type for the service worker: only required for `inline: true`.
+   */
+  type: WorkerType
+  /**
+   * Returns the corresponding script tag if `shouldRegisterSW` returns `true`.
+   */
+  toScriptTag: () => string | undefined
+}
+
 export interface VitePluginPWAAPI {
   /**
    * Is the plugin disabled?
    */
   disabled: boolean
+  /**
+   * Running on dev server?
+   */
+  pwaInDevEnvironment: boolean
+  /**
+   * Returns the PWA webmanifest url for the manifest link:
+   * <link rel="manifest" href="<webManifestUrl>" />
+   *
+   * Will also return if the manifest will require credentials:
+   * <link rel="manifest" href="<webManifestUrl>" crossorigin="use-credentials" />
+   */
+  webManifestData(): WebManifestData | undefined
+  /**
+   * How the service worker is being registered in the application.
+   *
+   * This option will help some integrations to inject the corresponding script in the head.
+   */
+  registerSWData(): RegisterSWData | undefined
   extendManifestEntries(fn: ExtendManifestEntriesHook): void
   /*
    * Explicitly generate the manifests.
    */
-  generateBundle(bundle?: OutputBundle): OutputBundle
+  generateBundle(bundle?: OutputBundle): OutputBundle | undefined
   /*
    * Explicitly generate the PWA services worker.
    */
@@ -245,3 +352,62 @@ export interface VitePluginPWAAPI {
 }
 
 export type ExtendManifestEntriesHook = (manifestEntries: (string | ManifestEntry)[]) => (string | ManifestEntry)[] | undefined
+
+/**
+ * Development options.
+ */
+export interface DevOptions {
+  /**
+   * Should the service worker be available on development?.
+   *
+   * @default false
+   */
+  enabled?: boolean
+  /**
+   * The service worker type.
+   *
+   * @default 'classic'
+   */
+  type?: WorkerType
+  /**
+   * This option will enable you to not use the `runtimeConfig` configured on `workbox.runtimeConfig` plugin option.
+   *
+   * **WARNING**: this option will only be used when using `generateSW` strategy.
+   *
+   * @default false
+   */
+  disableRuntimeConfig?: boolean
+  /**
+   * This option will allow you to configure the `navigateFallback` when using `registerRoute` for `offline` support:
+   * configure here the corresponding `url`, for example `navigateFallback: 'index.html'`.
+   *
+   * **WARNING**: this option will only be used when using `injectManifest` strategy.
+   */
+  navigateFallback?: string
+
+  /**
+   * This option will allow you to configure the `navigateFallbackAllowlist`: new option from version `v0.12.4`.
+   *
+   * Since we need at least the entry point in the service worker's precache manifest, we don't want the rest of the assets to be intercepted by the service worker.
+   *
+   * If you configure this option, the plugin will use it instead the default.
+   *
+   * **WARNING**: this option will only be used when using `generateSW` strategy.
+   *
+   * @default [/^\/$/]
+   */
+  navigateFallbackAllowlist?: RegExp[]
+
+  /**
+   * On dev mode the `manifest.webmanifest` file can be on other path.
+   *
+   * For example, **SvelteKit** will request `/_app/manifest.webmanifest`, when `webmanifest` added to the output bundle, **SvelteKit** will copy it to the `/_app/` folder.
+   *
+   * **WARNING**: this option will only be used when using `generateSW` strategy.
+   *
+   * @default `${vite.base}${pwaOptions.manifestFilename}`
+   * @deprecated This option has been deprecated from version `v0.12.4`, the plugin will use navigateFallbackAllowlist instead.
+   * @see navigateFallbackAllowlist
+   */
+  webManifestUrl?: string
+}
